DS-5372 Complete
An analytic cookie is set and a consent notice is displayed.

There was an idea that we will maintain a cross-subdomain cookie for the user's consent status, that would be shared with Vertex. You confirm in one application and the other does or doesn't show the consent banner based upon the user already having approved it.
For now, this idea is shelved because of the effort to change Vertex cookie logic and the effort needed to change the MkDocs Material template's use of a non-standard cookie that really isn't a "cookie".